Update unit tests to run on a docker container
Currently, this repo's unit tests run directly on the developer's
machine, or in CI on the Github Action's container. Since we'll be
deploying stela on Docker containers, and running it locally on Docker
containers, it is preferrable to instead run tests on Docker containers
as well. This commit updates the repo to do that.