Add Github Action to test then build and publish images
Upon a commit to main (or as needed on a feature branch) we want to
build a docker image that can be used to start containers running this
application and publish that image to ECR. This commit adds a Github
Action to achieve that. Since generally only builds from main will be
deployed, this action runs automatically on updates to main, and can be
run manually on feature branches. Since no build should ever be created
if tests are failing, this workflow first runs tests, and builds only if
they pass. Since the build workflow will run tests on main, the
test-only workflow has been disabled on main.