Add deployment to dev environment
This commit adds terraform configuration-as-code defining an EKS cluster
in AWS, creating on that cluster a deployment running two pods of stela,
and exposing stela's endpoints to the Internet via an application load
balancer. It also adds workflows to CI to validate the terraform code
and to deploy to dev upon merge to main (or a manual trigger).