Add POST endpoint for creating directives
In pursuit of our goal of adding directives to the Permanent app, this
commit adds a POST endpoint for creating a directive (and an associated
directive trigger). As the first non-health-check endpoint in this repo,
this endpoint establishes a few practices that merit mentioning. The
first is using Joi for input validation (we also use Joi in
notification-service). Also introduced is a `fixtures` repo, for storing
SQL scripts that generate seed data for use in testing. Finally, for
ease of development this commit alters tests to create their test
database on the `devenv` postgres container, and to copy the schema for
that test database from that container's `permanent` database. This is
necessary in order to get changes from migrations.