Add an endpoint to trigger directives admin directives for an account
The first trigger type we plan to release for directives is the admin
trigger type, meaning the directives are manually triggered by a
Permanent employee upon receiving word that it is time for the account's
legacy plan to be activated. This commit adds an endpoint that, given an
account ID, finds and executes all directives for that account with the
"admin" trigger type.