Add update endpoint for directives
Directives need to be updateable, so we need an endpoint to update them.
This commit adds one. It also updates this repo to handle all IDs as
strings, even if they happen to consist of numeric digits. This is for
two reasons. First, making the type of a variable `number` should
communicate that it represents a mathematical quantity which either
represents a real world quantity or will be involved in arithmetic.
IDs don't have these qualities, they are just uniquely identifying
*strings* of digits. Second, our numeric IDs are bigints in PostgreSQL,
and Javascript doesn't natively support integers that large (64 bits).
Thus handling IDs as strings avoids cumbersome and potenially
error-prone operations to convert PostgreSQL bigints to Javascript
numbers.