Stop depending local pg_dump version (to discuss)
I ran into a problem where my local pg_dump was an older version than
the one in the containers, so I couldn't load the database and tests
failed. It seems to me that it shouldn't matter what version I'm on in
the host.

This change works, but `devenv_database_1` seems like the wrong name
to use - shouldn't just "database" work? It didn't for me, but I could
be missing something. Will `devenv_database_1` always be the name of
the container?