Add prod deployments
We want to use `stela` for real, so it needs to be deployed to prod.
This commit sets up the prod deploy workflow. It also adds an "all test
environments" deploy workflow, to handle cases where we need to update
parts of the test cluster that our targeted deploys to dev and staging
won't affect.