Send notification email to new legacy contacts
We want to notify people when they become a Permanent user's legacy
contact. This commit updates the create endpoint for legacy contacts to
send a notification email to the new legacy contact, and updates the
update endpoint for legacy contacts to send the same notification email
if the update changed the email of the legacy contact.