Add GET /account/signup endpoint to get signup details
Currently, there is no way for the frontend to know the conditions under
which a user signed up for their Permanent account after the signup
takes place. We would like the frontend to show different welcome
screens to users who came to Permanent through different paths, so the
frontend needs this information. This commit adds a GET endpoint at
/account/signup, which returns the invite token the user signed up with
(which could have come from an actual invite, or from a promo page, or
be a placeholder value for users who came without an invite through the
standard signup flow).