Add admin endpoint to regenerate folder thumbnails
Our previous method of generating thumbnail images for folders on the
backend did not produce suitable images, so it isn't actually used.
Recent work has updated this process to generate better images that the
frontend will be able to use instead of selecting thumbnails itself.
Before the frontend can do this, though, we need to regenerate the
thumbnails of folders that predate this change. To that end, this commit
adds an admin-authenticated endpoint that will requeue for thumbnail
generation every folder created prior to a provided timestamp.