Make the API server an individual app under the stela repo
Now that we're adding lambdas to stela, it is becoming a monorepo
hosting multiple distinct applications (which may share some pieces). To
make way for this new structure, the API server that so far has been all
of stela is moved in this commit into it's own directory under the new
apps/ directory, where it will soon be joined by the metadata
processing lambda.