Add endpoints to manage featured archives
Currently the featured archives of the public archive gallery are stored
in a JSON file in web-app. We want to make these editable through our
admin portal, so we need endpoints to add and remove featured archives,
and to retrieve the current set of featured archives. This commit adds
these endpoints.