Fix test coverage analysis
When codecov was installed in this repo, the tests weren't yet running
in their own docker container. When the tests were moved to a container,
the coverage report wasn't sent from that container back to the Github
actions box where the codecov action runs, so coverage reports stopped
reaching codecov. This commit fixes this by voluming the coverage
directory on the test container.