Stop using ts-node
ts-node is not widely used, particularly in production. While working on
adding a metadata processing lambda to this project, it proved very
difficult to get ts-node to work with npm-workspaces. In the interest of
using more standard tools, this commit removes ts-node in favor of
running the compiled output of the TypeScript compiler (tsc). It adds
tscp to handle moving sql files into the compiled project, and tsc-watch
to handle hotloading in the dev environment.