Add POST /event endpoint
To avoid to tight a coupling to Mixpanel, we want our own event tracking
endpoint that records events in our database and, if desired, forwards
corresponding events to Mixpanel. This commit adds that endpoint. It
also refactors the authentication middleware to allow for endpoints that
can be authenticated as either user or admin.