Add middleware to optionally extract email from auth token
Our new GET /record email should work for both logged in users with a
JWT and for unauthenticated callers accessing public records. To that
end, we need a new middleware function that extracts an email from a JWT
if one is provided, but doesn't fail in the absence of a valid JWT. This
commit adds this middleware.

PER-9431: Make record/get v2 take and return multiple records

Co-authored-by: Natalie Martin <natalie@permanent.org>
Co-authored-by: Liam Lloyd <liam@permanent.org>