Add script to clean up the Archivematica dashboard
Once we start processing files with Archivematica, we will need to
periodically clean out the Archivematica dashboard, to prevent it from
being unusably slow if we ever log in to debug something or update
configuration. This commit adds a script that cleans out all completed
jobs from the dashboard. A future commit will set up this script to run
on a Kubernetes CronJob workload.