Added Archivematica cleanup cron job to deployment
Once we start using Archivematica for file processing, we will need a
cron process to periodically delete successfully completed jobs from the
Archivematica dashboard. This commit creates a Dockerfile to run the
Archivematica cleanup script, and updates the terraform configuration to
create a Kubernetes CronJob to run that docker image each hour.