Parse list of IPs properly in `extractIp`
The back-end will pass a list of proxies from the X-Forwarded-For
header, if needed[1]. The `extractIp` middleware was not written
to expect this format, and so would cause validation errors when passed
a list of proxy IPs. Update this middleware function to handle a proxy
IP list by extracting out the client IP from the list.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-For#syntax

PER-9651: "Publish archive" does not get checked off