Add account_space_update lambda handler
Since uploads tend to come in large bursts with spans of low activity in
between, we want to move asynchronous upload handling to lambdas, so we
can scale our resources according to present necessity. To that end,
this commit adds a package to stela containing a lambda handler that
will update the amount of storage an account has left whenever a new
file is stored in our systems.