Fix passing of SSL public keys into docker image
Currently, we try to pass the public SSL keys needed to connect to our
databases to the stela API docker image via a build variable, but this
build variable is defined too early in the file for it to be accessible
to the deployment image. This commit moves it later so it becomes
accessible to that image. It also adds the -e flag to the echo command
when we write the keys to a file, so newlines are preserved.