Keep 3 failed cron pods for debugging
Currently, our cron pods are deleted 10 seconds after the job completes.
This is good for conserving resources, but bad for debugging
(particularly when one of the things in need of debugging is logs going
to Sentry, as is presently the case). This commit improves the situation
by removing the time-to-live for completed cron pods, but instructing
kubernetes to keep up to 3 failed pods around so we can connect to them
and see what happens, while discarding successful pods as soon as the
job completes.