Remove failed cron pods after an hour
A recent commit set up our Archivematica cleanup cron to maintain up to
three failed pods for debugging. Apparently this is three pods *per
run*. In order to prevent the cluster from getting cluttered with a
bunch of pods when the cron breaks, this commit adds back a time-to-live
of 1 hour, so error pods will only stick around until the next run.