Add a test that user authentication middleware output passes validation
Recently, we updated our user authentication middleware to add the
subject encoded in the JWT to the request body. This caused most
authenticated endpoints to return 400 errors, because their request
validation didn't expect the presence of the subject. To avoid such
things happening in the future, this commit adds a middleware test that
checks that the request will pass a validation function that checks only
for the output of the auth token. A future commit will update all
validators for authenticated endpoints to share code with this
validator, so that a future fix to it would apply to all validators of
authenticated endpoints.

PER-9772: Optionally use `userSubjectFromAuthToken`