Expect user subject on all user authenticated requests
Recently, we updated our user authentication middleware to add the
subject encoded in the JWT to the request body. This commit updates
tests of all user-authenticated endpoints to expect that subject to be
added to the request. It also centralizes the validation rules for
fields added from the JWT into the shared validators module, to ensure
that future changes to fields from the JWT accurately impact tests.

PER-9772: Optionally use `userSubjectFromAuthToken`