Add lambda handler for attaching thumbnails to records
Now that Archivematica is generating thumbnails for us, we want to use
those instead of generating our own. This commit adds a lambda handler
that will be subscribed to S3 events fired when Archivematica generates
new .jpg files inside DIPs. The handler checks whether the .jpg is a
thumbnail, and if so creates a CDN URL for it and adds that URL, and the
S3 location of the thumbnail, to the corresponding record in our
database.