Add infrastructure as code for record thumbnail lambda
Because upload volume is highly variable, we would like to use a lambda
to attach thumbnails created by Archivematica to new records in our
database. This commit adds the necessary terraform configuration to
deploy such a lambda, and updates the deploy actions of the repo to
include the lambda in deploys.