Fix broken tests in Archivematica cleanup cron
Tests for the Archivematica cleanup cron are currently broken, because
the cleanup function is stored in the index.ts file, meaning the tests
must import the index.ts file. Because this file is a script, it runs
upon loading, interferring with the tests. It also causes the tests to
fail due to lack of environment variables (which aren't actually needed
due to mocking) in CI. To fix this problem, this commit moves the
cleanup function into a service.ts file, and tests that file rather than
index.ts.