Add thumbnail refresh script
Presently, we refresh expiring or soon-to-expire CDN URLs for thumbnails
when we retrieve the item the thumbnail belongs to. This assigns too
many responsibilities to our read endpoints and slows them down with
extra database calls. Instead, thumbnail refreshes should be handled by
a separate process responsible only for that. This commit introduces a
script that updates thumbnail URLs for all records and folders where
they will expire within the month, intended to be run as a weekly cron.