Add API design for global feature flags endpoints
We want to start releasing some features behind feature flags, and to
streamline this process we want to be able to manage feature flags from
our admin panel. To that end, we need endpoints we can use to view and
manage feature flags. This commit adds the OpenAPI specification for the
create, read, and update endpoints for feature flags.