Add email to request body when verifying user or admin authentication
Presently, the verifyUserOrAdminAuthentication middleware, used by the
POST /event endpoint, adds only the subject from the auth token to the
request body. We've been asked to include emails in Mixpanel events, so
this commit updates the middleware to add the email from the auth token
to the request body as well.