Update tests to work with changes to devenvs docker compose
When running in Github Actions, this repo's tests use the devenv repo's
docker compose to set up the database. Recent changes to the development
environment have caused that docker compose to expect the devenv
direction to have a .env file, which is currently not present when this
repo runs tests. This commit adds a step to the test action to create an
empty .env file to prevent docker from erroring when it sets up the
database.