Extract CDN URL creation function into a helper package
Both the thumbnail lambda and the access copy lambda need to be able to
generate CDN URLs. To avoid code reuse, this commit extracts a function
for generating such URLs into a package for S3-related helper functions.
It also adds an optional fileName argument, for use in the generation of
download URLs.