Extract function for getting fileIds from Archivematica package paths
Both the record thumbnail lambda and the access copy lambda extract
fileIds from the paths to Archivematica packages (and their contents) in
S3. To avoid duplicated code, this commit creates an archivematica-utils
package and adds this function to that package (in the future I expect
this package to hold a lot of metadata parsing code).