Add access copy lambda
In order to use Archivematica-generated access copies, we need to
represent them in our database. This commit adds a lambda handler, to be
triggered by the creation of an access copy in S3 by Archivematica,
which creates a file entry in our database corresponding to that file,
and links it to the appropriate record and parent file.