Add function to publish to SNS, with attributes
We'd like to start sending events reported to the /event endpoint to
SNS, so they can trigger async processes. These will come in one at a
time, so this commit adds a function to publish a single message to SNS
(implemented in terms of the existing function to publish multiple
messages). Also, because these events will be directed to the
appropriate SQS queues based on message attributes, this commit updates
the SNS publishing functions to accept messages with attributes.