Send events to SNS
We'd like to be able to trigger asynchronous processes based on events
fired through our /event API. To that end, this commit forwards such
events to an SNS topic. The intent is that processes interested in these
events will listen to an SQS queue subscribed to this topic, and
filtering the events it gets based on the Entity and Action attributes
of those events.