Add endpoints for third party developers to future API design
We want to create a smooth process for third-party developers to
integrate their apps with Permanent, or build their entire apps on top
of Permanent. In pursuit of this, this commit adds several endpoints to
our future API design, covering the creation of third-party developer
accounts, the creation and management of third party applications in
FusionAuth, and login via OAuth.