Add admin endpoint to delete orphaned folders
A recently fixed bug caused certain folders to delete only superficially
when users attempted to delete them. As a result, the users did not get
their storage back, and the contents of those folders can still appear
in searches. This commit adds an admin endpoint to re-trigger deletion
for all such folders to correct this situation.