Fix misleading Date types
There are a handful of interfaces in this repo that claim some of their
fields are Dates, when in fact they are strings. Because the code never
actually expects them to be dates, this doesn't cause errors, but it is
confusing. This commit updates the types of such fields to string.