Add GET /archive/{archiveId}/folders/shared endpoint
The new sharing menu wants to offer users the opportunity to move the
item they're sharing into an existing shared folder. This commit adds a
new endpoint that will retrieve all an archive's shared folders to
suppor this use case. Note that it retrieves only top-level shared
folders, the idea being the the user would navigate into their children
if they wanted to move the item to a descendent folder.