Optimize get folder and record queries
Currently, the queries for getting folders and records assemble the
lists of shares for all folders and records before joining the relevant
ones to the folders and records in question. This commit optimizes the
queries to only get the shares for the relevant folders and records in
the first place.