Add /archive/:archiveId/backfill-ledger admin endpoint
We have some archives in the system where storage wasn't historically
tracked via the application, but rather by hand, in order to facilitate
free storage for nonprofit partners. We can now achieve this while still
tracking storage in the application, and this in fact preferable for
partners because it allows them to track their storage balances more
easily. This commit adds an admin tool that can help migrate archives
from the old accounting system to the new one by backfilling the storage
ledger for all items in the archive that weren't originally included
therein.