Reorganize docs and commit docs site
We would like to make our API documentation available via a website
hosted on Github Pages. To that end, this commit adds the site output by
redocly to the repo. Also, to make the docs more useful, it unifies the
"present" and "future" docs into a single body, noting at the endpoint
level which are not yet implemented.