Add archive and share link to GET /records response
It is worthwhile for a record object to include basic information about
the archive the record lives in, so this commit adds an archive object
to the record object, including the archive's id, archive number, and
name. Note that the former two fields are already part of the record as
top-level fields; these are now deprecated and will be removed in the
future.

This commit also adds a shareLink object to records, including
information about the account that created the share link. This will
appear in the response only when a share token is used to access the
record(s).