Add archive name and share link to GET /folders response
In some cases we want to display the name of the archive where a folder
lives, so this commit adds a name field to the folder's archive object.

This commit also adds a shareLink object to folders, including
information about the account that created the share link. This will
appear in the response only when a share token is used to access the
folder(s).