Deploy docs to `gh-pages` branch
We want to serve our documentation via github pages, but we don't want
that compiled documentation to be stored in our main code branch.  This
sets up a workflow to push rendered documentation to the magical
`gh-pages` branch every time there is an update to `main`.

We'll need to configure GitHub Pages to look to that branch instead of
main.

Issue #291 Switch to a gh-pages branch for documentation