Remove shareLink field from records and folders
The shareLink field added to records and folders turned out not to be
useful (the information contained therein needed to be available in
cases where it couldn't easily be obtained from a record or folder
request). Thus, this commit removes it. Note that this is a breaking
change, which we wouldn't normally make, but the field was added
recently and no client has ever used it, so it is safe to remove in this
case.