Make type guards for objects with status or statusCode fields typesafe
Currently, when we check that an error has a status or statusCode field,
the code that does the checking briefly assumes that it does have such a
field. This could cause us to throw less helpful errors about underlying
errors not having this fields rather than throwing the underlying error.
Thus this commit updates these checks to safely verify whether these
fields exist.