Memoize `introspectAccessToken`
There are a few memoize packages out there on npm; this one has
typescript support and appears to be more popular than `memoize`.  It's
also a standalone package (as opposed to the lodash / emotion memoize
packages which seem to be just smaller parts of bigger libraries, and
also seem to be more focused on front-end contexts.

I chose a 10 second cache which translates to at most 6 calls to
FusionAuth per client, which I believe is *well* within their expected
rates.  We were seeing rate limits when running SFTP which was resulting
in closer to hundreds of calls per second.

Using 10 seconds is also low enough that a token that has been manually
invalidated on FusionAuth would stop working within 10 seconds, which
I am proposing does not reflect a meaningful security concern.

The biggest risk of this addition is potential memory leaks -- if somehow the
`setTimeout` for cache clearing is dropped or doesn't run.  That said,
the event cycle is managed by node and is pretty darn core to the way
the node runtime works.  If we're facing dropped timeout calls we're
probably looking at deeper issues.

Issue #314 Preventing FusionAuth rate limits