Don't delete build on start in local API server
Presently, when we build the docker container for the local stela API,
we build the project during the container build, but then when we start
the container we delete the API dist folder and rebuild it, but
incompletely, causing the server to fail to start. This was presumably
introduced in recent dependency upgrades. This commit removes the step
that deletes api/dist from the start script; this step has been
unnecessary for some time.