Log error if record_thumbnail_attacher doesn't update any rows
We're seeing cases in production where files are successfully processed
by archivematica, and SQS messages about the existence of thumbnails are
successfully proceesed by the record_thumbnail_attacher, but the records
don't actually get thumbnails. It seems likely that this is coming from
cases where, for some reason, the update query in the
record_thumbnail_attacher doesn't update any rows. This commit adds logs
for that case so we can see whether it's happening.