Create a dependabot group for Sentry updates
Dependabot recently tried to update one of the sentry packages we use,
and its PR was broken because this created conflicts with other sentry
packages that it declined to update. This commit seeks to avoid that
problem in the future by instructing dependabot to update all sentry
packages at once.