Fix middleware that fails on undefined requests
Before switching to the express.json() middleware, empty requests bodies
were represented by empty objects, meaning middleware that added things
to the request body was always able to do so. Now, however, empty
request bodies are undefined, meaning such middleware fails. This commit
adds an extra piece of middleware to run after express.json() and set
request.body to {} if it is undefined, allowing the rest of the
middleware to run successfully. Note that this is a stopgap, we should
eventually update all the middleware not to modify the request body.