Avoid using -target in terraform commands
Currently, we use -target when deploying to dev and staging to specify
whether to deploy the dev or staging workloads. This is not a
recommended use for the -target flag, and it means that in cases
where a deploy updates underlying infrastructure our dev and staging
deploy workflows don't work as intended and we have to use the full test
deploy workflow. This commit updates the dev and staging deploys to take
a map of images to update in the deployed workloads and use the presence
of values in this map to determine which workloads need updated. This
allows us to stop using the -target flag and to delete the
now-unnecessary full test deploy workflow.