Update GET /record to return records in the order they were requested
Currently, the order in which GET /record returns records is undefined.
This remains the API contract, but as a hotfix this commit updates the
endpoint to return them in the order they're listed in the request,
because that is what our web app currently expects. This change should
be reverted when the web app no longer relies on this behavior.