Add metadata attacher lambda
The last major piece of Archivematica-related infrastructure is a way to
add metadata extracted from files by Archivematica to the Permanent
database. This commit adds a lambda to do so. For the moment it only
focuses on image metadata and has only been tested on JPGs; adding
support for other image types and videos is planned to follow quickly.