Remove redundant dependabot configuration
We were defining the npm package ecosystem for each separate stela
package, but I have not seen this in other monorepo projects that use
Dependabot.  The documumentation is not entirely clear, but this
comment[1] in a related issue suggests our approach is the correct one.

Issue #587 Use directories in dependabot config to avoid repetition

1. https://github.com/dependabot/dependabot-core/issues/4993#issuecomment-1289133027